# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Die krasse Raufaser API",
        "description": "Mega Doku für krasse Sachen",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/save-gold": {
            "get": {
                "tags": [
                    "SaveData"
                ],
                "summary": "Retrieve saved data",
                "description": "Returns all saved data",
                "operationId": "getSavedData",
                "responses": {
                    "200": {
                        "description": "Data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SaveData"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "SaveData"
                ],
                "summary": "Save data (deprecated - do not use)",
                "description": "Stores data in the database",
                "operationId": "saveData",
                "requestBody": {
                    "description": "Pass data to be saved",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "charName",
                                    "transactionType"
                                ],
                                "properties": {
                                    "charName": {
                                        "type": "string",
                                        "example": "Hero123"
                                    },
                                    "transactionType": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "AmountGold": {
                                        "type": "string",
                                        "example": "1000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data saved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/upload-transactions": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Upload transactions JSON file",
                "description": "Uploads transactions from a JSON file and saves them to the database",
                "operationId": "uploadTransactions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "profileKeys": {
                                        "type": "object"
                                    },
                                    "profiles": {
                                        "properties": {
                                            "Xyssa - Blutkessel": {
                                                "properties": {
                                                    "totalMoney": {
                                                        "type": "integer",
                                                        "example": 1656634395
                                                    },
                                                    "transactions": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "1": {
                                                                    "type": "string",
                                                                    "example": "Player - Item - Amount"
                                                                },
                                                                "2": {
                                                                    "type": "string",
                                                                    "example": "Player - Item - Amount"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "MoneyTransactions": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "TimeStamp": {
                                                                    "type": "integer",
                                                                    "example": 1739705784
                                                                },
                                                                "PlayerGUID": {
                                                                    "type": "string",
                                                                    "example": "Player-578-0ABE9596"
                                                                },
                                                                "Action": {
                                                                    "type": "string",
                                                                    "example": "Deposit"
                                                                },
                                                                "Value": {
                                                                    "type": "integer",
                                                                    "example": 1000
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transactions uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transactions uploaded successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid JSON format"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions",
                "description": "Retrieve a list of all stored transactions",
                "operationId": "getTransactions",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "player_name": {
                                                "type": "string",
                                                "example": "Xyssa - Blutkessel"
                                            },
                                            "amount": {
                                                "type": "integer",
                                                "example": 1000
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "Deposit"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Player-578-0ABE9596"
                                            },
                                            "transaction_timestamp": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-02-16 11:36:24"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "SaveData": {
                "title": "SaveData",
                "description": "SaveData model schema",
                "required": [
                    "charName",
                    "transactionType"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "charName": {
                        "type": "string",
                        "example": "Hero123"
                    },
                    "transactionType": {
                        "type": "integer",
                        "example": 1
                    },
                    "AmountGold": {
                        "type": "string",
                        "example": "1000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "SaveData",
            "description": "SaveData"
        },
        {
            "name": "Transactions",
            "description": "Transactions"
        }
    ]
}